fastapi1/requirements.txt
uvicorn==0.34.0
fastapi==0.115.8
tinydb
pandas
openai
load_dotenv

